+++ benchmark_model_distributed/benchmarkers/direct_multi_thread_benchmarker.py
@@ -1,5 +1,3 @@
-import datetime
-import numpy as np
 import threading
 
 from time import perf_counter
